core/requirements.txt
asgiref==3.7.2
Django==3.2.5
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
